Refactor worker logic into service class
